CE_4201/requirements.txt
paho-mqtt
smbus
numpy
imusensor
pygame
gTTS
gpiozero
tinydb
opencv-python
onnxruntime
mediapipe
ollama
ai-edge-litert
tensorflow
keras-tuner